+++ src/GameTetris.cpp
@@ -519,7 +519,7 @@ void GameTetris::EndLevel()
 
 	level_end_animation_end_tick_ =
 		tick_ +
-		Fixed16RoundToInt(g_fixed16_one + sound_player_.GetMelodyDuration(music_id)) * GameInterface::c_update_frequency;
+		uint32_t(Fixed16RoundToInt(g_fixed16_one + sound_player_.GetMelodyDuration(music_id))) * GameInterface::c_update_frequency;
 }
 
 void GameTetris::NextLevel()
